testsuite: Remove empty files

+++ testsuite/testsuite.cpp
@@ -962,13 +962,17 @@ int download(const std::string & source, const std::string& referer,
                 curl_easy_setopt(handle, CURLOPT_HTTPHEADER, nullptr);
                 curl_slist_free_all(hlist);
             }
+            auto written = ftell(fp);
             fclose(fp);
 
-            if(!error) {
+            if (!error && written > 0) {
                 std::cout << " DONE\n";
-            }
-            else {
-                std::cout << " CURL Error " << error << std::endl;
+            } else {
+                if (error) {
+                    std::cout << " CURL Error " << error << std::endl;
+                } else {
+                    std::cout << " Empty file\n";
+                }
                 unlink(dest.c_str());
                 dest = "";
                 return -1;
